requirements.txt
PyMuPDF
python-docx
numpy
faiss
pytest
openai
faiss-cpu
python-dotenv
sentence-transformers
cryptography
